Desktop/comp90024_team_61-main/backend/fission/functions/sentiment_trends/requirements.txt
elasticsearch8==8.14.0
pandas==2.2.2
python-dateutil==2.9.0